pyproject.toml
[project]
name = "langchain-tutorials-campusx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.18",
    "langchain-google>=0.1.1",
    "langchain-google-genai>=2.0.9",
    "langchain-huggingface>=0.1.2",
    "langchain-openai>=0.3.5",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.6.1",
    "streamlit>=1.42.2",
]
